refactored browse section into breeds section, applied to dogs and cats
