+++ src/components/cards-and-sections/styles/BrowseSection.module.css
@@ -10,6 +10,30 @@
   margin: 1rem 0;
 }
 
+.info {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.info p {
+  font-size: var(--fs-subheading);
+  margin: 0 2rem;
+  margin-bottom: 10%;
+  text-align: center;
+}
+
+.pagination {
+  display: flex;
+  justify-content: center;
+  gap: 2rem;
+  margin: 2rem 0;
+}
+
+.pagination button {
+  color: var(--clr-accent);
+}
+
 @media screen and (min-width: 1000px) {
   .cards {
     gap: 3rem;
