+++ src/data/navData.js
@@ -3,7 +3,7 @@ const NavData = [
   { text: 'Search Animals', to: '/search' },
   { text: 'Search Organisations', to: '/organisations' },
   { text: 'Dog Breeds', to: '/dog-breeds/1' },
-  { text: 'Cat Breeds', to: '/cat-breeds' },
+  { text: 'Cat Breeds', to: '/cat-breeds/1' },
 ];
 
 export default NavData;
