+++ src/index.js
@@ -24,7 +24,7 @@ root.render(
           <Route path="/search" element={<SearchAnimalsPage />} />
           <Route path="/organisations" element={<SearchOrgsPage />} />
           <Route path="/dog-breeds/:page" element={<DogBreedsPage />} />
-          <Route path="/cat-breeds" element={<CatBreedsPage />} />
+          <Route path="/cat-breeds/:page" element={<CatBreedsPage />} />
           <Route path="/results/:location/:page/:sort" element={<ResultsPage />} />
           <Route path="/results/:location/:page/:sort/:filters" element={<ResultsPage />} />
           <Route path="/animal/:id" element={<PetViewPage />} />
