+++ src/pages/cat-breeds/_CatBreedsPage.js
@@ -1,25 +1,15 @@
 import React from 'react';
 
 import CatBreedsHero from './CatBreedsHero';
-import BrowseSection from '../../components/cards-and-sections/BrowseSection';
-
-import placeholder from '../../assets/images/breeds/cat-hero.jpg';
+import BreedSection from '../../components/cards-and-sections/BreedSection';
 
 function _CatBreedsPage() {
   const heading = 'Browse by type';
-  const breeds = [
-    { name: 'Test', image: placeholder, available: '31,212' },
-    { name: 'Test', image: placeholder, available: '31,212' },
-    { name: 'Test', image: placeholder, available: '31,212' },
-    { name: 'Test', image: placeholder, available: '31,212' },
-    { name: 'Test', image: placeholder, available: '31,212' },
-    { name: 'Test', image: placeholder, available: '31,212' },
-  ];
 
   return (
     <main>
       <CatBreedsHero />
-      <BrowseSection heading={heading} items={breeds} />
+      <BreedSection heading={heading} />
     </main>
   );
 }
