--- src/pages/dog-breeds/styles/DogBreedsPage.module.css
@@ -1,23 +1,0 @@
-.info {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.info p {
-  font-size: var(--fs-subheading);
-  margin: 0 2rem;
-  margin-bottom: 10%;
-  text-align: center;
-}
-
-.pagination {
-  display: flex;
-  justify-content: center;
-  gap: 2rem;
-  margin: 2rem 0;
-}
-
-.pagination button {
-  color: var(--clr-accent);
-}
