+++ src/pages/results/styles/ResultsCard.module.css
@@ -20,7 +20,7 @@
   font-size: var(--fs-subheading);
   color: var(--clr-accent);
   margin: 0.5rem 0;
-  line-height: 1;
+  line-height: 1.3;
   overflow: hidden;
   text-overflow: ellipsis;
   display: -webkit-box;
